Refactors analyzers into a trait and extracts analysis loop.
The Validator trait replaces individual invocations of is_haiku()
and similar methods and adds an accept() method to replace the
ad-hoc lines-between-x-and-y decisions at the callers.

The validators now return errors as a Vec of ClassifyError, an
enum type that's easier to count and sort (and in the future,
render in HTML). This enables a simple loop to find the
'BestInterpretation', that which has the fewest errors.

Only the terminal output is affected for now.