Creates a "Shelf" of Dictionaries, to allow querying over all of them.
This change makes several big, simplifying assumptions around how the
dictionaries are (mostly not) merged. The important thing now is to
allow adding unknown entries easily; this is more important than
overriding existing entries, and the trickyness happens with the latter.

Alternatives:

I first tried to create a separate CompositeDictionary object to proxy
the requests, but got stuck into a mess of either invalid signatures
(in the one-object-per-request case) and self-references/lifetimes
(in the shared object case). This avoids all that for a bit.