Fills out the /analyze page with the best interpretation.
The page now has a left-right split, with the user input on the left
and an analysis of the prose on the right, color coded for missing
words and with analysis/errors/warnings at the bottom.

For expediency, this was done with lots of raw html formatting in
code. If and when this is updated, it should be switched to use
inner templates.